changed time type to sizet

+++ src/time.c
@@ -6,26 +6,26 @@
 /*   By: gyong-si <gyongsi@student.42.fr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/30 12:27:54 by gyong-si          #+#    #+#             */
-/*   Updated: 2024/04/08 12:26:13 by gyong-si         ###   ########.fr       */
+/*   Updated: 2024/04/09 14:27:18 by gyong-si         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/philo.h"
 
-void usleep_ms(uint32_t milliseconds)
+void	usleep_ms(size_t milliseconds)
 {
-	uint32_t	t;
+	size_t	t;
 
 	t = get_current_time();
 	while (1)
 	{
 		if (get_current_time() - t >= milliseconds)
-			break;
+			break ;
 		usleep(50);
 	}
 }
 
-uint32_t	get_current_time(void)
+size_t	get_current_time(void)
 {
 	struct timeval	time;
